fix: normalize car id handling to string

+++ app/src/main/java/com/fx/zfcar/car/ReportActivity.kt
@@ -153,7 +153,7 @@ class ReportActivity : AppCompatActivity(), View.OnClickListener {
             } else if (reportAdapter.type == ReportAdapter.ReportType.OFFLINE) {
                 val offlineItem = (reportAdapter.getItem(position) as ReportItem.OfflineItem).data
                 val intent = Intent(this, OfflineDetailActivity::class.java)
-                intent.putExtra(OfflineDetailActivity.KEY_CAR_ID, offlineItem.carId.toInt())
+                intent.putExtra(OfflineDetailActivity.KEY_CAR_ID, offlineItem.carId)
                 intent.putExtra(OfflineDetailActivity.KEY_CAR_NUM, offlineItem.carNum)
                 intent.putExtra(OfflineDetailActivity.KEY_START, binding.tvStartDate.text.toString())
                 intent.putExtra(OfflineDetailActivity.KEY_END, binding.tvEndDate.text.toString())
@@ -161,7 +161,7 @@ class ReportActivity : AppCompatActivity(), View.OnClickListener {
             } else if (reportAdapter.type == ReportAdapter.ReportType.EXPIRED) {
                 val expireItem = (reportAdapter.getItem(position) as ReportItem.ExpiredItem).data
                 val intent = Intent(this, CarInfoActivity::class.java)
-                intent.putExtra(CarInfoActivity.KEY_CAR_ID, expireItem.carId.toInt())
+                intent.putExtra(CarInfoActivity.KEY_CAR_ID, expireItem.carId)
                 startActivity(intent)
             }
         }
